Added many new features to effects.
Status effects can now be made to rely on having a certain number of stacks from an arbitrary number of other effects (including itself) before applying its damage/recursive effects.  Status effects can now be set to only apply once and then remove itself. Status effects that are required to apply the main effect can be set to be removed when the main effect is applied.